refactor: fix surface handling
