+++ packages/tools/src/tools/displayTools/Contour/contourDisplay.ts
@@ -82,8 +82,7 @@ async function render(
     contourData = await computeAndAddRepresentation(
       segmentationId,
       Representations.Contour,
-      () => polySeg.computeContourData(segmentationId, { viewport }),
-      () => undefined
+      () => polySeg.computeContourData(segmentationId, { viewport })
     );
 
     polySegConversionInProgressForViewportId.set(viewport.id, false);
@@ -273,7 +272,20 @@ function _checkContourNormalsMatchViewport(
   return true;
 }
 
+/**
+ * Default function to call when segmentation representation is updated
+ *
+ * @param viewport
+ * @returns
+ */
+function getUpdateFunction(
+  viewport: Types.IVolumeViewport | Types.IStackViewport
+): (segmentationId: string) => Promise<void> {
+  return null;
+}
+
 export default {
+  getUpdateFunction,
   render,
   removeRepresentation,
 };
