+++ packages/tools/src/utilities/segmentation/computeAndAddRepresentation.ts
@@ -1,6 +1,5 @@
 import type { SegmentationRepresentations } from '../../enums';
 import addRepresentationData from '../../stateManagement/segmentation/internalAddRepresentationData';
-import { addSegmentationListener } from './segmentationEventManager';
 
 /**
  * Computes a segmentation representation and subscribes to future segmentation updates.
@@ -11,13 +10,11 @@ import { addSegmentationListener } from './segmentationEventManager';
  * @param updateFunction - Optional function to update UI/state on segmentation change.
  * @param onComputationComplete - Optional callback invoked after computation completes.
  * @returns - A promise that resolves with the computed representation data.
-
  */
 async function computeAndAddRepresentation<T>(
   segmentationId: string,
   type: SegmentationRepresentations,
   computeFunction: () => Promise<T>,
-  updateFunction?: (segmentationId: string) => void,
   onComputationComplete?: () => void
 ): Promise<T> {
   // Compute the specific representation data
@@ -30,9 +27,6 @@ async function computeAndAddRepresentation<T>(
   });
 
   onComputationComplete?.();
-
-  // Subscribe to any changes in the segmentation data for real-time updates
-  addSegmentationListener(segmentationId, type, updateFunction);
 
   return data;
 }
